--- src/main/java/gift/web/exception/CategoryNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.web.exception;
-
-public class CategoryNotFoundException extends RuntimeException{
-    public CategoryNotFoundException(String message) {
-        super(message);
-    }
-}
