--- src/main/java/gift/web/exception/MemberNotFoundException.java
@@ -1,5 +1,0 @@
-package gift.web.exception;
-
-public class MemberNotFoundException extends RuntimeException {
-    public MemberNotFoundException(String message) { super(message); }
-}
