--- src/main/java/gift/web/exception/OptionNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.web.exception;
-
-public class OptionNotFoundException extends RuntimeException {
-    public OptionNotFoundException(String message) {
-        super(message);
-    }
-}
