--- src/main/java/gift/web/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.web.exception;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
