--- src/main/java/gift/web/exception/WishProductNotFoundException.java
@@ -1,8 +1,0 @@
-package gift.web.exception;
-
-public class WishProductNotFoundException extends RuntimeException{
-
-    public WishProductNotFoundException(String message) {
-        super(message);
-    }
-}
