+++ src/main/java/gift/web/exception/notfound/CategoryNotFoundException.java
@@ -1,0 +1,7 @@
+package gift.web.exception.notfound;
+
+public class CategoryNotFoundException extends NotFoundException{
+    public CategoryNotFoundException() {
+        super("카테고리가 없슴다!");
+    }
+}
