+++ src/main/java/gift/web/exception/notfound/OptionNotFoundException.java
@@ -1,0 +1,7 @@
+package gift.web.exception.notfound;
+
+public class OptionNotFoundException extends NotFoundException {
+    public OptionNotFoundException() {
+        super("옵션이 없슴다!");
+    }
+}
