+++ src/main/java/gift/web/exception/notfound/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package gift.web.exception.notfound;
+
+public class ProductNotFoundException extends NotFoundException {
+    public ProductNotFoundException() {
+        super("상품이 없슴다!");
+    }
+}
