+++ src/main/java/gift/web/exception/notfound/WishProductNotFoundException.java
@@ -1,0 +1,8 @@
+package gift.web.exception.notfound;
+
+public class WishProductNotFoundException extends NotFoundException{
+
+    public WishProductNotFoundException() {
+        super("위시 제품이 없슴다!");
+    }
+}
